get rid of CellMapper interface
