--- cell/src/main/java/jetbrains/jetpad/cell/mappers/CellMapper.java
@@ -1,27 +1,0 @@
-/*
- * Copyright 2012-2015 JetBrains s.r.o
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package jetbrains.jetpad.cell.mappers;
-
-import jetbrains.jetpad.cell.Cell;
-import jetbrains.jetpad.cell.toUtil.HasCounters;
-import jetbrains.jetpad.model.event.EventHandler;
-import jetbrains.jetpad.model.property.PropertyChangeEvent;
-import jetbrains.jetpad.values.Color;
-
-public interface CellMapper extends HasCounters, EventHandler<PropertyChangeEvent<Cell>> {
-  void setAncestorBackground(Color background);
-  void refreshProperties();
-}
